[Experimental] More work on getting the NN to operate properly. I am not convinced the genomes are being randomly selected properly. Also we are not getting any attempts to generate input for the game. The NN does generate the next generation, but does not iterate through it properly. During the second generation, the NN will skip to a random species and increment at random intervals, or not increment the current species at all and just generate a bunch of new genomes for the active species, without changing the generation. I am going to comb through the MarI/O very carefully and comment our code to explain what is happening, as well as changing most of the QVectors to QMaps in order to reflect as much of the Lua functionality in C++ as possible.

+++ RoboCopClient/genome.h
@@ -52,7 +52,7 @@ public:
 
     bool ContainsLink(Gene link);
     bool SameSpecies(const Genome &other, int innovationSize);
-    int Disjoint(const QVector<Gene*> &other, int innovationSize);
+    float Disjoint(const QVector<Gene*> &other, int innovationSize);
     float Weights(const QVector<Gene*> &other, int innovationSize);
